requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.11.14
aiosignal==1.3.2
async-timeout==5.0.1
attrs==25.3.0
cffi==1.17.1
click==8.1.8
discord.py==2.5.2
frozenlist==1.5.0
idna==3.10
joblib==1.4.2
multidict==6.2.0
nltk==3.9.1
propcache==0.3.0
pycparser==2.22
PyNaCl==1.5.0
python-dotenv==1.1.0
regex==2024.11.6
tqdm==4.67.1
typing_extensions==4.12.2
yarl==1.18.3
